update README
Update README with confirmed meaning of 'expires', and with python snippet from github.com/ncssar/sartopo_python.
Confirmed with Matt that 'expires' does not refer to subscription expiration date, it refers to any arbitrary time (in milliseconds) in the future after which you want this request to become invalid, as another security feature.  Matt suggests 2 minutes in the future.